chore: 기한 만료 체험 시 UI 차별화 CSS 추가
