+++ src/app/myReservation/style.module.css
@@ -27,6 +27,7 @@
 }
 
 .reservationBox {
+  position: relative;
   -webkit-dispaly: flex;
   display: flex;
   margin-bottom: 24px;
@@ -41,6 +42,21 @@
     margin-bottom: 0;
   }
 
+  .overlay {
+    position: absolute;
+    top: 0;
+    left: 0;
+    bottom: 0;
+    right: 0;
+    background: var(--black);
+    opacity: 0.1;
+    z-index: 1;
+
+    & ~ * {
+      color: var(--gray) !important;
+    }
+  }
+
   .thumbnail {
     position: relative;
     flex: 1;
@@ -51,6 +67,10 @@
       content: '';
       display: block;
       padding-bottom: 100%;
+    }
+
+    .overlay {
+      opacity: 0.4;
     }
   }
 
@@ -124,4 +144,8 @@
       color: var(--gray);
     }
   }
+}
+
+.reservationBox.noClick {
+  cursor: default;
 }
